+++ R/geocode_evavelo.R
@@ -47,7 +47,6 @@ geocode_evavelo <- function(data){
 find_wrong_geocoding <- function(data){
   if(attr(data, "geocoded") == FALSE)
     stop("data must be geocoded before getting geocoding errors")
-
   geocoded_fields <- c(ville_heb = "ville_heb",
                        iti_depart_itineraire = "iti_depart_itineraire",
                        iti_arrivee_itineraire = "iti_arrivee_itineraire",
@@ -57,12 +56,17 @@ find_wrong_geocoding <- function(data){
   ## Check tables communes (with cog)
   tables_communes <- data$table_communes %>%
     dplyr::filter(!is.na(.data$nom_commune) & is.na(.data$latitude)) %>%
-    dplyr::select(.data$nom_commune, .data$cog)
+    dplyr::select(
+      .data$nom_commune, .data$cog,
+      dplyr::any_of(dplyr::starts_with("proposition_"))
+    )
 
   ## Check ville_res (with cp)
-  ville_res <- find_wrong_cities("ville_res",
-                                 data$enquete,
-                                 fields_to_keep = c("id_quest", "pays_res", "cp_res"))
+  ville_res <- find_wrong_cities(
+    "ville_res",
+    data$enquete,
+    fields_to_keep = c("id_quest", "pays_res", "cp_res", "proposition_cp_res")
+  )
 
   return(
     c(list(tables_communes = tables_communes,
@@ -74,19 +78,48 @@ find_wrong_geocoding <- function(data){
 
 }
 
-#' Find cities that couldn't be geocoded by geocode_cities()
+#' Find cities that couldn't be geocoded by geocode_cities() or that have been interpreted due to a mispelling
 #'
 #' @param var_name string corresponding to the city name variable to check
 #' @param enquete data.frame containing the var_name variable and an associated *_cog
 #' @param fields_to_keep fields that will be exported along with var_name
+#' @param include_interpretation Should we include geocoding interpretation? if FALSE only errors will be output.
 #'
 #' @return a data.frame with cities that couldn't be geocoded
 #' @keywords internal
 
-find_wrong_cities <- function(var_name, enquete, fields_to_keep = "id_quest"){
+find_wrong_cities <- function(var_name,
+                              enquete,
+                              fields_to_keep = "id_quest",
+                              include_interpretation = TRUE){
+
+  ## Keeping only necessary information
   lon <- paste0(var_name, "_lon")
-  enquete %>%
-    dplyr::filter(!is.na(get(var_name)) & is.na(get(lon))) %>%
-    dplyr::select(dplyr::all_of(fields_to_keep),
-           dplyr::all_of(var_name))
+  propos <- paste0("proposition_", var_name)
+  fields_to_keep <- c(fields_to_keep, var_name, propos, lon)
+
+  enquete <- enquete %>%
+    dplyr::select(dplyr::any_of(fields_to_keep)) %>%
+    dplyr::filter(!is.na(get(var_name))) %>%
+    dplyr::relocate(
+      dplyr::starts_with("proposition"),
+      .after = dplyr::last_col()
+    )
+
+  ## Geocoding errors (no longitude)
+  result <- enquete %>%
+    dplyr::filter(
+      is.na(get(lon))
+    )
+
+  ## Geocoding interpretation
+  if(include_interpretation & propos %in% names(enquete)){
+    result <- enquete %>%
+      dplyr::filter(!is.na(get(propos)) & !is.na(get(lon))) %>%
+      dplyr::bind_rows(result, .)
+  }
+
+  result %>%
+    dplyr::select(- !!lon)
+
 }