+++ tests/testthat/test-geocode_cities_cp.R
@@ -23,10 +23,32 @@ test_that("geocode_cities_cp works", {
     "Canberra", "22000", 2, 2.0, "Australie"
   )
 
+  nomatim_country_df <- tibble::tribble(
+    ~city, ~postcode, ~a, ~b, ~country,
+    "Nantes", "44000", 1, 1.0, "France",
+    "Canberra", "22000", 2, 2.0, "Australie",
+    "Yamba", NA, 3, 3.0, "Australia" ## Small city can only be found using Nomatim
+  )
+
+
   na_df <- tibble::tribble(
     ~city, ~postcode, ~a, ~b, ~country,
     NA, NA, 1, 1.0, "France",
     NA, NA, 2, 2.0, "Australie"
+  )
+
+  mispelling_country_df <- tibble::tribble(
+    ~city, ~postcode, ~a, ~b, ~country,
+    "Nantes", "44000", 1, 1.0, "France",
+    "Canberra", NA, 2, 2.0, "Australie",
+    "sidney", "22000", 2, 2.0, "Australie"
+  )
+
+  mispelling_all_df <- tibble::tribble(
+    ~city, ~postcode, ~a, ~b, ~country,
+    "Nante", "44000", 1, 1.0, "France",
+    "Canberra", NA, 2, 2.0, "Australie",
+    "sidney", "22000", 2, 2.0, "Australie"
   )
 
   expected_name <- c(names(correct_df), "city_lat", "city_lon", "city_cog")
@@ -64,8 +86,24 @@ test_that("geocode_cities_cp works", {
                       country_col = country) %>%
       suppressMessages())
 
+  expect_snapshot_output(
+    geocode_cities_cp(nomatim_country_df,
+                      city_col = city,
+                      cp_col = postcode,
+                      country_col = country) %>%
+      suppressMessages())
+
+  expect_snapshot_output(
+    geocode_cities_cp(mispelling_all_df,
+                      city_col = city,
+                      cp_col = postcode ,
+                      country_col = country) %>%
+      suppressMessages())
+
+
+  ## Test name of results (which differs if city/cp are correct or not)
   expect_equal(
-    geocode_cities_cp(wrong_cp_df,
+    geocode_cities_cp(correct_df,
                       city_col = city,
                       cp_col = postcode ,
                       country_col = country) %>%
@@ -73,6 +111,50 @@ test_that("geocode_cities_cp works", {
       names(),
     expected_name)
 
+  expect_equal(
+    geocode_cities_cp(wrong_cp_df,
+                      city_col = city,
+                      cp_col = postcode ,
+                      country_col = country) %>%
+      suppressMessages() %>%
+      names(),
+    c(expected_name, "proposition_city", "proposition_postcode")
+  )
+
+  expect_equal(
+    geocode_cities_cp(mispelling_country_df,
+                      city_col = city,
+                      cp_col = postcode ,
+                      country_col = country) %>%
+      suppressMessages() %>%
+      names(),
+    c(expected_name, "proposition_city")
+  )
+
 })
 
+test_that("geocode_df_foreign_cities works", {
+
+  df <- tibble::tribble(
+    ~city, ~postcode, ~a, ~b, ~country,
+    "Nantes", "44000", 1, 1.0, "France",
+    "Canberra", "22000", 2, 2.0, "Australie",
+    "Yamba", NA, 3, 3.0, "Australia" ## Small city can only be found using Nomatim
+  )
+  ## Expect no message when geocoding
+
+  expect_message({
+    geo_df <- geocode_df_foreign_cities(
+      df,
+      city_col = city,
+      country_col = country
+    )},
+    regexp = NA
+  )
+
+  expect_snapshot_output(
+    geo_df
+  )
+
+})